backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
requests>=2.31.0
numpy>=1.26.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyttsx3>=2.90
# Optional dependencies (install for voice support):
# vosk>=0.3.45
# sentence-transformers>=2.2.2
# chromadb>=0.4.18
